we connected the API to get forecasts
I am giving you the coordinates after I get the initial response, then you go and get the forecast. Then we go into the function getForecast and create the apiURL which is a onecall api, which coordinates and unit and apiKey. We then tell the axios: can you go get the this apiURL and once you get the response, go displayForecast:
You see we console.log the response.